requirements.txt
opencv-python==4.1.0.25
dnfal==0.1.0
numpy==1.16.0
gpustat==0.6.0
requests==2.22.0
cvtlib==0.1.0
psutil==5.6.3
Django==2.2.4
PyJWT==1.7.1
djangorestframework==3.10.2
celery==4.3.0
torch==1.3.0
